+++ tests/integration/flex_api/v1/test_insights_segments.py
@@ -65,3 +65,91 @@ class InsightsSegmentsTestCase(IntegrationTestCase):
         actual = self.client.flex_api.v1.insights_segments("segment_id").fetch()
 
         self.assertIsNotNone(actual)
+
+    def test_list_request(self):
+        self.holodeck.mock(Response(500, ''))
+
+        with self.assertRaises(TwilioException):
+            self.client.flex_api.v1.insights_segments.list(token="token")
+
+        headers = {'Token': "token", }
+        self.holodeck.assert_has_request(Request(
+            'get',
+            'https://flex-api.twilio.com/v1/Insights/Segments',
+            headers=headers,
+        ))
+
+    def test_read_empty_response(self):
+        self.holodeck.mock(Response(
+            200,
+            '''
+            {
+                "segments": [],
+                "meta": {
+                    "page": 0,
+                    "page_size": 50,
+                    "first_page_url": "https://flex-api.twilio.com/v1/Insights/Segments?ReservationId=WReeeff42d5326cfdaa0f7a1181d00f1b1&PageSize=50&Page=0",
+                    "previous_page_url": null,
+                    "url": "https://flex-api.twilio.com/v1/Insights/Segments?ReservationId=WReeeff42d5326cfdaa0f7a1181d00f1b1&PageSize=50&Page=0",
+                    "next_page_url": null,
+                    "key": "segments"
+                }
+            }
+            '''
+        ))
+
+        actual = self.client.flex_api.v1.insights_segments.list()
+
+        self.assertIsNotNone(actual)
+
+    def test_read_full_response(self):
+        self.holodeck.mock(Response(
+            200,
+            '''
+            {
+                "segments": [
+                    {
+                        "segment_id": "02dd8a65-c9e3-550b-9e98-5b11abb9b3a5",
+                        "external_id": "EV4ab3c50aeeb9dc48c96f7290cf626a37",
+                        "queue": "Everyone",
+                        "external_contact": "+13051515226",
+                        "external_segment_link_id": null,
+                        "date": "2022-02-08T08:00:00Z",
+                        "account_id": "76e4c1c3-6f93-5d08-ae66-aad4b9092f2d",
+                        "external_segment_link": null,
+                        "agent_id": "0e4a8059-18a3-5e48-9efb-29ae3213e7fe",
+                        "agent_phone": "+13051515226",
+                        "agent_name": "Flex Insights",
+                        "agent_team_name": "Everyone",
+                        "agent_team_name_in_hierarchy": "Everyone",
+                        "agent_link": null,
+                        "customer_phone": "+420606078007",
+                        "customer_name": "+420606078007",
+                        "customer_link": null,
+                        "segment_recording_offset": null,
+                        "media": null,
+                        "assessment_type": [
+                            null
+                        ],
+                        "assessment_percentage": [
+                            null
+                        ],
+                        "url": "https://flex-api.twilio.com/v1/Insights/Segments/02dd8a65-c9e3-550b-9e98-5b11abb9b3a5"
+                    }
+                ],
+                "meta": {
+                    "page": 0,
+                    "page_size": 50,
+                    "first_page_url": "https://flex-api.twilio.com/v1/Insights/Segments?ReservationId=WReeeff42d5326cfdaa0f7a1181d00f1b1&PageSize=50&Page=0",
+                    "previous_page_url": null,
+                    "url": "https://flex-api.twilio.com/v1/Insights/Segments?ReservationId=WReeeff42d5326cfdaa0f7a1181d00f1b1&PageSize=50&Page=0",
+                    "next_page_url": null,
+                    "key": "segments"
+                }
+            }
+            '''
+        ))
+
+        actual = self.client.flex_api.v1.insights_segments.list()
+
+        self.assertIsNotNone(actual)