+++ tests/integration/microvisor/v1/test_account_config.py
@@ -143,6 +143,37 @@ class AccountConfigTestCase(IntegrationTestCase):
 
         self.assertIsNotNone(actual)
 
+    def test_update_request(self):
+        self.holodeck.mock(Response(500, ''))
+
+        with self.assertRaises(TwilioException):
+            self.client.microvisor.v1.account_configs("key").update(value="value")
+
+        values = {'Value': "value", }
+
+        self.holodeck.assert_has_request(Request(
+            'post',
+            'https://microvisor.twilio.com/v1/Configs/key',
+            data=values,
+        ))
+
+    def test_update_response(self):
+        self.holodeck.mock(Response(
+            200,
+            '''
+            {
+                "key": "first",
+                "value": "place",
+                "date_updated": "2021-01-01T12:34:56Z",
+                "url": "https://microvisor.twilio.com/v1/Configs/first"
+            }
+            '''
+        ))
+
+        actual = self.client.microvisor.v1.account_configs("key").update(value="value")
+
+        self.assertIsNotNone(actual)
+
     def test_delete_request(self):
         self.holodeck.mock(Response(500, ''))