+++ tests/integration/microvisor/v1/test_account_secret.py
@@ -139,6 +139,36 @@ class AccountSecretTestCase(IntegrationTestCase):
 
         self.assertIsNotNone(actual)
 
+    def test_update_request(self):
+        self.holodeck.mock(Response(500, ''))
+
+        with self.assertRaises(TwilioException):
+            self.client.microvisor.v1.account_secrets("key").update(value="value")
+
+        values = {'Value': "value", }
+
+        self.holodeck.assert_has_request(Request(
+            'post',
+            'https://microvisor.twilio.com/v1/Secrets/key',
+            data=values,
+        ))
+
+    def test_update_response(self):
+        self.holodeck.mock(Response(
+            200,
+            '''
+            {
+                "key": "first",
+                "date_rotated": "2021-01-01T12:34:56Z",
+                "url": "https://microvisor.twilio.com/v1/Secrets/first"
+            }
+            '''
+        ))
+
+        actual = self.client.microvisor.v1.account_secrets("key").update(value="value")
+
+        self.assertIsNotNone(actual)
+
     def test_delete_request(self):
         self.holodeck.mock(Response(500, ''))