+++ twilio/rest/flex_api/v1/insights_questionnaires_question.py
@@ -330,6 +330,7 @@ class InsightsQuestionnairesQuestionInstance(InstanceResource):
             'answer_set_id': payload.get('answer_set_id'),
             'allow_na': payload.get('allow_na'),
             'usage': deserialize.integer(payload.get('usage')),
+            'answer_set': payload.get('answer_set'),
             'url': payload.get('url'),
         }
 
@@ -416,6 +417,14 @@ class InsightsQuestionnairesQuestionInstance(InstanceResource):
         :rtype: unicode
         """
         return self._properties['usage']
+
+    @property
+    def answer_set(self):
+        """
+        :returns: Question's Answer set
+        :rtype: dict
+        """
+        return self._properties['answer_set']
 
     @property
     def url(self):